Refactor approach to use different protocol instead of parameter.
This is a new approach to switch between http and http batching.
It only is different in this regard from the previous attempts,
and only contains refactorings besides this change.
