+++ src/pkg/egress/syslog/writer_factory_test.go
@@ -43,7 +43,7 @@ var _ = Describe("EgressFactory", func() {
 
 	Context("when the url begins with https and enables batching", func() {
 		It("returns an single https writer", func() {
-			url, err := url.Parse("https://syslog.example.com?batching=true")
+			url, err := url.Parse("https-batch://syslog.example.com")
 			Expect(err).ToNot(HaveOccurred())
 			urlBinding := &syslog.URLBinding{
 				URL: url,
@@ -55,7 +55,7 @@ var _ = Describe("EgressFactory", func() {
 			retryWriter, ok := writer.(*syslog.RetryWriter)
 			Expect(ok).To(BeTrue())
 
-			_, ok = retryWriter.Writer.(*syslog.BatchHTTPSWriter)
+			_, ok = retryWriter.Writer.(*syslog.HTTPSBatchWriter)
 			Expect(ok).To(BeTrue())
 		})
 	})
